changes to conftest and test-grin-client and test-conversion
